Separate the controller with the control panel. Enable multiple controllers interact with the panel.
