+++ PIDControl/PID4FIWARE.py
@@ -8,9 +8,8 @@ Updated on Tue Mar 15 15:47:00 2022
 
 @author: jdu
 
-PID controller with Fiware interface & Web-based controller panel
+PID controller with Fiware interface
 """
-
 import time
 import os
 from simple_pid import PID
@@ -18,8 +17,6 @@ import requests
 from filip.clients.ngsi_v2 import ContextBrokerClient
 from filip.models.base import FiwareHeader
 from filip.models.ngsi_v2.context import NamedCommand, ContextEntity, NamedContextAttribute
-import threading
-import PySimpleGUIWeb as sg
 from keycloak_token_handler.keycloak_python import KeycloakPython
 
 
@@ -28,9 +25,8 @@ class Control:
 
     def __init__(self):
         """Initialization"""
-        # use envirnment variables  
-        os.environ["CONFIG_FILE"] = "False"
-
+        # # use envirnment variables
+        # os.environ["CONFIG_FILE"] = "False"
         # define parameters
         self.params = {}
         self.params['name'] = os.getenv("NAME", 'PID_1')
@@ -191,109 +187,7 @@ class Control:
             os.abort()
 
 
-class ControllerPanel:
-    def __init__(self):
-        # initialize controller parameters (in dict)
-        self.params = self.initialize_params()
-
-        # FIWARE parameters
-        self.cb_url = os.getenv("CB_URL", "http://localhost:1026")
-        self.entity_id = os.getenv("NAME", 'PID_1')
-        self.entity_type = "PID_Controller"
-        self.service = os.getenv("FIWARE_SERVICE", '')
-        self.service_path = os.getenv("FIWARE_SERVICE_PATH", '')
-
-        # Create the fiware header
-        fiware_header = FiwareHeader(service=self.service, service_path=self.service_path)
-
-        # Create orion context broker client
-        self.ORION_CB = ContextBrokerClient(url=self.cb_url, fiware_header=fiware_header)
-
-        # initialize gui window
-        sg.theme("DarkBlue")
-        param_bars = [
-            [sg.Text(param.capitalize(), size=(10, 1)), sg.InputText(self.params[param], key=param)]
-            for param in self.params.keys()
-        ]
-        buttons_bar = [[sg.Button("Send"), sg.Button("Read")]]
-        layout = param_bars + buttons_bar
-        self.window = sg.Window("PID controller", layout, web_port=80, web_start_browser=True)
-
-    def gui_update(self):
-        """Update the shown text on web GUI"""
-        for param in self.params.keys():
-            self.window[param].update(self.params[param])
-
-    def gui_loop(self):
-        """GUI main loop"""
-        try:
-            while True:
-                event, values = self.window.read(timeout=10)
-
-                if event in (sg.WINDOW_CLOSED, None):
-                    break
-                elif event == "Send":
-                    self.send(values)
-                elif event == "Read":
-                    print("Read", flush=True)
-                    self.read()
-        finally:
-            print("panel loop fails")
-            self.window.close()
-            os.abort()
-
-    def read(self):
-        """Read parameter values from context broker"""
-        try:
-            params_update = self.initialize_params()
-            for param in self.params.keys():
-                params_update[param] = float(self.ORION_CB.get_attribute_value(entity_id=self.entity_id,
-                                                                               entity_type=self.entity_type,
-                                                                               attr_name=param))
-            self.params = params_update
-        except requests.exceptions.HTTPError as err:
-            msg = err.args[0]
-            if "NOT FOUND" not in msg.upper():
-                raise
-            print("Cannot find controller entity")
-            self.params = self.initialize_params()
-        finally:
-            self.gui_update()
-
-    def send(self, params):
-        """Send new parameter values to context broker"""
-        for param in self.params.keys():
-            try:
-                value = float(params[param])
-                self.ORION_CB.update_attribute_value(entity_id=self.entity_id,
-                                                     entity_type=self.entity_type,
-                                                     attr_name=param,
-                                                     value=value)
-            except ValueError:
-                print(f"Wrong value type of {param}: {params[param]}. Must be numeric!")
-
-    @staticmethod
-    def initialize_params():
-        """Initialize the values of all control parameters"""
-        # initialize controller parameters shown on panel
-        params = {
-            "Kp": "Proportional gain",
-            "Ki": "Integral gain",
-            "Kd": "Derivative gain",
-            "lim_low": "Lower limit of output",
-            "lim_upper": "Upper limit of output",
-            "setpoint": "The set point of control variable"
-        }
-        return params
-
-
 if __name__ == "__main__":
     pid_controller = Control()
     pid_controller.create_entity()
-    panel = ControllerPanel()
-
-    # Parallelism with multi thread
-    th1 = threading.Thread(target=pid_controller.control_loop, daemon=False)
-    th2 = threading.Thread(target=panel.gui_loop, daemon=False)
-    th1.start()
-    th2.start()
+    pid_controller.control_loop()
